Print pointers using iostream pointer print, don't try casting onto integer
git-svn-id: https://svn.code.sf.net/p/stlplus/stlplus3@507 ebc0ae9a-4168-4a02-ab27-d0bb227af65f

+++ trunk/tests/digraph_test/digraph_test.cpp
@@ -17,12 +17,12 @@ typedef stlplus::digraph<std::string,int> string_int_graph;
 
 std::ostream& operator<<(std::ostream& output, string_int_graph::iterator i)
 {
-  return output << ((long)((void*)&*i)) << "->" << *i;
+  return output << ((void*)&*i) << "->" << *i;
 }
 
 std::ostream& operator<<(std::ostream& output, string_int_graph::arc_iterator i)
 {
-  return output << ((long)((void*)&*i)) << "->" << *i;
+  return output << ((void*)&*i) << "->" << *i;
 }
 
 std::ostream& operator<<(std::ostream& output, string_int_graph::arc_vector arcs)
